+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add event on PVC if bound to update PV.
+
 ## [0.1.0]
 
 ### Added
 - Initial implementation
 
+[Unreleased]: https://github.com/piraeusdatastore/linstor-affinity-controller/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/piraeusdatastore/linstor-affinity-controller/releases/tag/v0.1.0